requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.4
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
